+++ api/src/modules/account/accounts.events.ts
@@ -2,7 +2,7 @@ import { InjectQueue } from '@nestjs/bull'
 import { Injectable } from '@nestjs/common'
 import { Queue } from 'bull'
 import { Jobs, Queues } from 'src/utils/constants'
-import IAccountEvents, { NewConfirmEmailRequestArgs } from './iaccount.events'
+import IAccountEvents, { NewConfirmEmailRequestArgs } from './i-accounts.events'
 
 @Injectable()
 export default class AccountEvents implements IAccountEvents {
