+++ api/src/modules/app.module.ts
@@ -2,10 +2,10 @@ import { BullModule } from '@nestjs/bull'
 import { Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
 import { MongooseModule } from '@nestjs/mongoose'
-import AccountsModule from './account/account.module'
+import AccountsModule from './account/accounts.module'
 import AuthModule from './auth/auth.module'
 import ProfilesModule from './profiles/profiles.module'
-import ResetPasswordModule from './reset_password/reset_password.module'
+import ResetPasswordModule from './reset_password/reset-password.module'
 import UsersModule from './users/users.module'
 
 const config = ConfigModule.forRoot()
