+++ api/src/modules/auth/auth.controller.ts
@@ -1,9 +1,9 @@
 import { Body, Controller, HttpException, HttpStatus, Inject, Post, Req, Res } from '@nestjs/common'
 import { Request, Response } from 'express'
 import { Routes, Services, refreshKey } from 'src/utils/constants'
-import { refreshTokenCookieOptions } from 'src/utils/cookieOptions'
-import IUserService from '../users/IUser.services'
-import IAuthService from './IAuth.services'
+import { refreshTokenCookieOptions } from 'src/utils/cookie-options'
+import IUserService from '../users/i-users.services'
+import IAuthService from './i-auth.services'
 import { LoginArgs } from './validators/login'
 import { LogoutArgs } from './validators/logout'
 import RegisterArgs from './validators/register'
