+++ api/src/modules/mail/i-mail.services.ts
