+++ api/src/modules/profiles/profiles.controllers.ts
@@ -2,10 +2,10 @@ import { Body, Controller, HttpException, Inject, Patch, UseGuards } from '@nest
 import AccountConfirmedGuard from 'src/guards/confirmed'
 import { RoleGuard } from 'src/guards/role'
 import { Roles, Routes, Services } from 'src/utils/constants'
-import IProfileService from './IProfiles.services'
-import UpdateEmailArgs from './validators/updateEmail'
-import UpdatePasswordArgs from './validators/updatePassword'
-import UpdateProfileArgs from './validators/updateProfile'
+import IProfileService from './i-profiles.services'
+import UpdateEmailArgs from './validators/update-email'
+import UpdatePasswordArgs from './validators/update-password'
+import UpdateProfileArgs from './validators/update-profile'
 
 @Controller(Routes.profiles)
 @UseGuards(AccountConfirmedGuard())
