+++ api/src/modules/profiles/profiles.events.ts
@@ -2,7 +2,7 @@ import { InjectQueue } from '@nestjs/bull'
 import { Injectable } from '@nestjs/common'
 import { Queue } from 'bull'
 import { Jobs, Queues } from 'src/utils/constants'
-import IProfileEvents, { NewEmailUpdatedArgs, NewPasswordUpdatedArgs } from './IProfiles.events'
+import IProfileEvents, { NewEmailUpdatedArgs, NewPasswordUpdatedArgs } from './i-profiles.events'
 
 @Injectable()
 export default class ProfileEvents implements IProfileEvents {
