+++ api/src/modules/refresh/refresh.repository.ts
@@ -3,7 +3,7 @@ import { InjectModel } from '@nestjs/mongoose'
 import { Model } from 'mongoose'
 import IMapper from 'src/common/mapper'
 import { Mappers, Models } from 'src/utils/constants'
-import IRefreshRepo from './IRefresh.repository'
+import IRefreshRepo from './i-refresh.repository'
 import Refresh from './refresh.domain'
 import RefreshDTO from './refresh.dto'
 import { RefreshDocument } from './refresh.model'
