+++ api/src/modules/refresh/refresh.services.ts
@@ -3,9 +3,9 @@ import * as crypto from 'crypto'
 import * as moment from 'moment'
 import IMapper from 'src/common/mapper'
 import { Mappers, Repos } from 'src/utils/constants'
-import generateId from 'src/utils/generateId'
-import IRefreshRepo from './IRefresh.repository'
-import IRefreshService from './IRefresh.services'
+import generateId from 'src/utils/generate-id'
+import IRefreshRepo from './i-refresh.repository'
+import IRefreshService from './i-refresh.services'
 import Refresh from './refresh.domain'
 import RefreshDTO from './refresh.dto'
