+++ api/src/modules/users/users.events.ts
@@ -2,7 +2,7 @@ import { InjectQueue } from '@nestjs/bull'
 import { Injectable } from '@nestjs/common'
 import { Queue } from 'bull'
 import { Jobs, Queues } from 'src/utils/constants'
-import IUserEvents, { NewUserExistArgs } from './IUsers.events'
+import IUserEvents, { NewUserExistArgs } from './i-users.events'
 
 @Injectable()
 export default class UserEvents implements IUserEvents {
