+++ api/src/modules/users/users.repository.ts
@@ -3,7 +3,7 @@ import { InjectModel } from '@nestjs/mongoose'
 import { Model } from 'mongoose'
 import IMapper from 'src/common/mapper'
 import { Mappers, Models } from 'src/utils/constants'
-import IUsersRepo from './IUsers.repository'
+import IUsersRepo from './i-users.repository'
 import User from './users.domain'
 import UserDTO from './users.dto'
 import { UserDocument } from './users.model'
