+++ api/src/modules/users/users.services.ts
@@ -3,10 +3,10 @@ import IMapper from 'src/common/mapper'
 import IHash from 'src/services/hashing'
 import { Events, Mappers, Repos, Roles, Services, defaultAvatar } from 'src/utils/constants'
 import { Left, Right } from 'src/utils/either'
-import generateId from 'src/utils/generateId'
-import IUserService, { CreateUserArgs, CreateUserResult, GetUserResult } from './IUser.services'
-import IUserEvents from './IUsers.events'
-import IUserRepo from './IUsers.repository'
+import generateId from 'src/utils/generate-id'
+import IUserEvents from './i-users.events'
+import IUserRepo from './i-users.repository'
+import IUserService, { CreateUserArgs, CreateUserResult, GetUserResult } from './i-users.services'
 import User from './users.domain'
 import UserDTO from './users.dto'
